+++ source/gauge_fixing.h
@@ -1,8 +1,8 @@
 #ifndef GAUGEFIXING_H
 #define GAUGEFIXING_H
 
-unsigned SU3_gauge_fix(double complex *U, const unsigned short config);
+unsigned SU3_gauge_fix(float complex *U, const unsigned short config);
 
-// void SU3_print_gaugefixed_U(double complex * U, double complex *U_aux, char filename[max_length_name]);
+// void SU3_print_gaugefixed_U(float complex * U, float complex *U_aux, char filename[max_length_name]);
 
 #endif
